Make Version::as_major_world fail on everything but correct versions

+++ wow_message_parser/src/ir_printer/mod.rs
@@ -90,7 +90,7 @@ pub enum IrWorldVersion {
         major: u8,
         minor: u8,
         patch: u8,
-        exact: u8,
+        exact: u16,
     },
 }
